adding 2 numbers together

+++ numder adding.py
@@ -2,8 +2,9 @@
 #12-09-14
 #numder adding
 
-numder1 = int(input("please enter numder: "))
-numder2 = int(input("please enter numder: "))
+numder1 = float(input("please enter numder: "))
+numder2 = float(input("please enter numder: "))
 
 answer = numder1 + numder2
 print(answer)
+
